Refactor ffprobe helpers
